fix(qlever): resolve compilation errors in stress/benchmark test files
Four pre-existing compilation bugs fixed:

- BenchmarkDataGenerator.java: wrong import java.util.IntStream → java.util.stream.IntStream
  (IntStream is in java.util.stream, not java.util)
- StressTestConfig.java: missing import java.util.Arrays (used for Arrays.stream/toString)
- QLeverBenchmarkTest.java:708: double→long lossy conversion; added explicit (long) cast
- QuickBenchmarkTest.java: testBasicQueryPerformance() and testDifferentQueryComplexities()
  missing throws Exception declaration (executeSelect throws SparqlEngineException)
- StressTestRunner.java: TimeoutException caught but never thrown; collapsed into
  InterruptedException check inside the generic catch block
- QLeverFfiExceptionTest.java: duplicate class declaration from "Quicksave" commit;
  removed the spurious second @DisplayName + class header at lines 35-42

All qlever stress/benchmark files now compile cleanly under --enable-preview --release 25.

+++ test/org/yawlfoundation/yawl/qlever/QLeverBenchmarkTest.java
@@ -701,7 +701,7 @@ class QLeverBenchmarkTest {
                 .average()
                 .orElse(0);
 
-            long meanSize = results.stream()
+            long meanSize = (long) results.stream()
                 .filter(r -> r.success)
                 .mapToLong(BenchmarkResult::resultCount)
                 .average()
